Use a shared_ptr to EntityVariable in Geometry to fix lifetime issues.
In Entity::geometryInFather we create such a shared_ptr and pass it to
the Geometry constrzuctor. This ensures that corners live as long as
the Geometry.
The existing constructors that take a reference to an EntityVariable
create a share_ptr with a noop deleter from the stack variable (using
Dune::stackobject_to_shared_ptr)