Also support ALUGrid for LookUpData
For ALUGrid this cartesian index mapper needs to be instantiated in a very special
way and one cannot create it just by passing a reference to the grid.
This creation is done in opm-simulators/ebos/eclalugridvanguard.hh. Then
the mapper can be accessed from the Vanguard and passed to LookupData.
For this one constructor now takes the mapper as an extra argument.

To still allow a construction from just the grid view, the class
stores a pointer to the mapper which is a nullptr if no mapper was
provided from the outside.

In addition change the constructors from taking a grid to taking a grid
view instead. From the grid view it is possible to access the
underlying grid, too.